Split recommendation product helpers
